+++ src/ExcelImport/ExcelImport.php
@@ -11,8 +11,6 @@ use ilObjUser;
 use ilSession;
 use ilSrUserEnrolmentPlugin;
 use ilUserDefinedFields;
-use PHPExcel;
-use PHPExcel_Shared_Date;
 use PhpOffice\PhpSpreadsheet\Shared\Date;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use srag\DIC\SrUserEnrolment\DICTrait;
@@ -175,7 +173,7 @@ class ExcelImport
         $rows = $excel->getSheetAsArray();
 
         /**
-         * @var Spreadsheet|PHPExcel $spreadsheet
+         * @var Spreadsheet $spreadsheet
          */
         $spreadsheet = Closure::bind(function ()/* : Spreadsheet*/ {
             return $this->workbook;
@@ -418,11 +416,7 @@ class ExcelImport
      */
     protected function handleSetPasswordFormatDateTime(stdClass &$user)/*: void*/
     {
-        if (self::version()->is54()) {
-            $date = Date::excelToDateTimeObject($user->{ExcelImportFormGUI::KEY_FIELDS}->{self::FIELDS_TYPE_ILIAS}->passwd__original_date_value);
-        } else {
-            $date = PHPExcel_Shared_Date::ExcelToPHPObject($user->{ExcelImportFormGUI::KEY_FIELDS}->{self::FIELDS_TYPE_ILIAS}->passwd__original_date_value);
-        }
+        $date = Date::excelToDateTimeObject($user->{ExcelImportFormGUI::KEY_FIELDS}->{self::FIELDS_TYPE_ILIAS}->passwd__original_date_value);
 
         // Modules/DataCollection/classes/Fields/Datetime/class.ilDclDatetimeRecordRepresentation.php::formatDate
         switch (self::dic()->user()->getDateFormat()) { // Assume date format for current user which has uploaded the excel file